+++ src/backend/core/models.py
@@ -168,8 +168,7 @@ class User(AbstractBaseUser, BaseModel, auth_models.PermissionsMixin):
     )
     language = models.CharField(
         max_length=10,
-        choices=settings.LANGUAGES,
-        default=settings.LANGUAGE_CODE,
+        blank=True,
         verbose_name=_("language"),
         help_text=_("The language in which the user wants to see the interface."),
     )